allow struct lifetime to vary with respect to 'de
Makes the Deserialize impl take an additional lifetime 'a (the
resulting Cow borrow lifetime) which is bounded above by 'de.
Enables the resulting Cow to have a shorter borrow than 'de -- this
is needed when embedding the types in a struct with derived Deserialize:

    #[derive(serde::Deserialize)]
    struct X<'x> {
        #[serde(borrow)]
        cow: CowStr<'x>
    }

Previously, this wouldn't compile, as the derive-generated Deserialize
impl looks like:

    impl Deserialize<'de: 'x, 'x> for X<'x> { /* ... */ }

But the CowStr lifetime wants to actually unify with 'de, so you get
compiler complaints that you should specify 'x: 'de.